Sat, Nov 4, 2023, 9:01 PM +07:00

+++ cs50w/week4/commerce/auctions/views.py
@@ -172,6 +172,8 @@ def listing(request):
 
 @login_required(login_url="login")
 def delete(request):
+    if request.method == "GET":
+        return HttpResponseRedirect()
     listing = get_object_or_404(Listing, id=request.GET.get("id"))
     if request.user == listing.user:
         listing.delete()
